% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rect_pack.R
\name{calc_small_box}
\alias{calc_small_box}
\title{Find a small box to store all the given rectangles}
\usage{
calc_small_box(
  rect_widths,
  rect_heights,
  aspect_ratios = c(1.61803, 1/1.61803),
  verbosity = 0L
)
}
\arguments{
\item{rect_widths, rect_heights}{widths and heights of the rectangles to pack.}

\item{aspect_ratios}{Vector of box aspect ratios to be tested. Aspect ratio 
is defined here as \code{width / height}. Default: \code{c(1.61803, 1/1.61803)}
i.e. golden ratio}

\item{verbosity}{Level of debugging output. Default: 0 (no output)}
}
\value{
list with w elements: box width and height
}
\description{
This is a brute force search with a simple heuristic, and is not 
guaranteed to find the box with the miinimum area, but simply a box
that snugly fits the rectangles without too much wasted space.
}
\examples{
# Find a minimal box to fit 10 random rectangles.
# Search for boxes with aspect ratios in seq(0.5, 2, length.out = 20)
set.seed(2)
N <- 10
widths  <- sample(N)
heights <- sample(N)
box <- calc_small_box(widths, heights, aspect_ratios = seq(0.5, 2, length.out = 20))
box
pack_rects(box$width, box$height, widths, heights)
}
